Manage LLDB ourselves instead of loading a script into it
This odyssey began when I couldn't call the command interpreter to run e.g.
`frame info` when we're using a script to drive LLDB's overall execution. I
believe this is because the command interpreter context is locked at the time of
the initial script, and then it can't be reused in a new context to run commands
from inside the overall script.

Although managing LLDB ourselves involved a few tedious steps, we also gain full
control over the debugger's execution, so it's likely the better path to take in
any case.